requirements.txt
boto3
python-dotenv
tqdm
pandas
transformers
torch
datasets
pillow
accelerate
python-pptx